batch line rendering

+++ Jungle_KMG_Week02/UObject.h
@@ -25,13 +25,13 @@ inline void* UObject::operator new(size_t size) {
 	void* ptr = ::operator new(size);
 	_allocationMap[ptr] = size;
 	++_totalCount;
-	_totalMemory += size;
+	_totalMemory += static_cast<uint32>(size);
 	return ptr;
 }
 
 inline void UObject::operator delete(void* p) {
 	--_totalCount;
-	_totalMemory -= _allocationMap[p];
+	_totalMemory -= static_cast<uint32>(_allocationMap[p]);
 	_allocationMap.erase(p);
 	free(p);
 }
